test: add test for escaping hex sequence round trip
- Add EscapeHexCellWriteHandlerRoundTripTest to verify hex escapes on write are read back literally
- Remove redundant HexEscapeRoundTripTest which tested similar functionality
- Import EscapeHexCellWriteHandler where necessary in write handler tests
- Ensure round trip correctness for cells containing hexadecimal escape patterns in data fields
